refactor: move schema to admin directory
